+++ factory/src/main/java/io/github/factoryfx/factory/storage/filesystem/FileSystemFactoryStorageHistory.java
@@ -13,7 +13,6 @@ import org.slf4j.LoggerFactory;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
-import java.nio.file.Paths;
 import java.util.*;
 import java.util.function.Consumer;
 import java.util.stream.Stream;
@@ -48,17 +47,9 @@ public class FileSystemFactoryStorageHistory<R extends FactoryBase<?, R>> {
     }
 
     public R getHistoryFactory(String id) {
-        StoredDataMetadata storedDataMetadata = null;
-        for (StoredDataMetadata metaData : getHistoryFactoryList()) {
-            if (metaData.id.equals(id)) {
-                storedDataMetadata = metaData;
-
-            }
-        }
-        if (storedDataMetadata == null) {
-            throw new IllegalStateException("cant find storedDataMetadata for factory: " + id + " in history");
-        }
-        return migrationManager.read(readFile(Paths.get(historyDirectory.toString() + "/" + id + ".json")), storedDataMetadata.dataStorageMetadataDictionary);
+        StoredDataMetadata metaData = migrationManager.readStoredFactoryMetadata(readFile(historyDirectory.resolve(id + "_metadata.json")), false);
+        updateCache(metaData);
+        return migrationManager.read(readFile(historyDirectory.resolve(id + ".json")), metaData.dataStorageMetadataDictionary);
     }
 
     private void visitHistoryFiles(Consumer<Path> visitor) {
@@ -69,16 +60,20 @@ public class FileSystemFactoryStorageHistory<R extends FactoryBase<?, R>> {
         }
     }
 
-    public Collection<StoredDataMetadata> getHistoryFactoryList() {
-        if (cache.isEmpty()) {
-            visitHistoryFiles(path -> {
-                if (path.toString().endsWith("_metadata.json")) {
-                    StoredDataMetadata storedDataMetadata = migrationManager.readStoredFactoryMetadata(readFile(path));
-                    cache.put(storedDataMetadata.id, storedDataMetadata);
-                }
-            });
+    public Collection<StoredDataMetadata> getHistoryFactoryList(boolean light) {
+        if (light && !cache.isEmpty()) {
+            return cache.values();
         }
-        return cache.values();
+
+        List<StoredDataMetadata> result = new ArrayList<>();
+        visitHistoryFiles(path -> {
+            if (path.toString().endsWith("_metadata.json")) {
+                StoredDataMetadata storedDataMetadata = migrationManager.readStoredFactoryMetadata(readFile(path), light);
+                result.add(storedDataMetadata);
+                cache.put(storedDataMetadata.id, StoredDataMetadata.createLightStoredDataMetadata(storedDataMetadata));
+            }
+        });
+        return result;
     }
 
     public void updateHistory(R factoryRoot, StoredDataMetadata metadata) {
@@ -86,8 +81,7 @@ public class FileSystemFactoryStorageHistory<R extends FactoryBase<?, R>> {
 
         writeFile(historyDirectory.resolve(id + ".json"), objectMapper.writeValueAsString(factoryRoot, OutputStyle.COMPACT));
         writeFile(historyDirectory.resolve(id + "_metadata.json"), objectMapper.writeValueAsString(metadata, OutputStyle.COMPACT));
-        cache.put(id, metadata);
-
+        updateCache(metadata);
         houseKeeping();
 
     }
@@ -134,9 +128,15 @@ public class FileSystemFactoryStorageHistory<R extends FactoryBase<?, R>> {
         });
     }
 
+    private void updateCache(StoredDataMetadata metadata) {
+        if (!cache.isEmpty()) {
+            cache.put(metadata.id, StoredDataMetadata.createLightStoredDataMetadata(metadata));
+        }
+    }
+
     private void houseKeeping() {
         if (maxConfigurationHistory == Integer.MAX_VALUE) {return;}
-        List<StoredDataMetadata> collect = getHistoryFactoryList().stream().toList();
+        List<StoredDataMetadata> collect = getHistoryFactoryList(true).stream().toList();
         int numToRemove = collect.size() - maxConfigurationHistory;
         if (numToRemove > 0) {
             collect.stream().sorted(Comparator.comparing(a -> a.creationTime)).limit(numToRemove).forEach(smd -> {