+++ factory/src/main/java/io/github/factoryfx/factory/storage/inmemory/InMemoryDataStorage.java
@@ -30,11 +30,14 @@ public class InMemoryDataStorage<R extends FactoryBase<?,R>> implements DataStor
     @Override
     public R getHistoryData(String id) {
         DataAndStoredMetadata<R> data = storage.get(id);
+        if (data == null) {
+            throw new RuntimeException("No history element found for id " + id);
+        }
         return data.root.utility().copy();
     }
 
     @Override
-    public Collection<StoredDataMetadata> getHistoryDataList() {
+    public Collection<StoredDataMetadata> getHistoryDataList(boolean light) {
         return storage.values().stream().map(item -> item.metadata).collect(Collectors.toList());
     }