+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 4.1.5
+* **DataStorage**
+  * add lightweight metadata for factory history (changeSummary == null && dataStorageMetadataDictionary == null)
+* **MicroserviceResource**
+  * getHistoryFactoryList returns factories with lightweight metadata by default
+
 # 4.1.4
 * **Microservice**
   * modify config revert message to show timestamp and comment instead of id